First commit - Organizing FR and US as separate repo.
The idea is that the user - (e.g., a private user, or
deux.commonaccord.org or source.commonaccord.org), mounts this repo in
their own instance, based on the GitHub name.  So this is mounted in
Doc/GH/CommonAccord/Form-Agt/.

This is experimental, and let’s see if it feels good.